Rename QueueGCManager
